Added a memory to the test setup
This will also be used when this project is running in an FPGA, whenever
that will be.

Could probably do with a testbench. The memory is accessed on the
falling clock edge. It is a sync RAM with byte strobes which are used
on writes, hopefully compatible with most FPGAs BRAM.