Initial go at ALU instructions
This is a little compllicated: the ALU is setup in the memory stage with
the result latched in the following register stage, where the result
is written into the register.

This needs a Mux for the register write data port since it can now be
sourced from either the databus or the ALU.

Also fixed a problem with halting.