Initial crack at M(rN)
Uses the ALU in the memorystage1; result is mux'd into the address.
Seems to work but needs more testing.

Also changed the halt mechanism to do the $fatal in the top level. Makes
it easier to set an arbitary time that the processor will run after
seeing the OPCODE_HALT in fetchstage0.