Added ALUMI - ALU op with 15 bit immediate
This uses a cycle signal shared by load and store so it can do its N(rM)
additions via the same mux.

The immediate is encoded a bit oddly: the low 12 bits at the bottom, and
the high 3 bits are where the cycle width/signed indicators would normally
be.

This leaves flow control as the last big part of the puzzle. But before
doing that I think I will remove some ALU ops, like inc and dec, which
are pointless with add using ALUMI.