Initial go at conditional branch/jump
Jump is untested. Branch seems to work, but there is a 3 clock branch
hazard, which means 3 nops need to be added to the tail of a branch.

The condition test is implemented as a combinational block.

Next up I want to move the condition codes into their own self-
contained regsiter.