Added support for Jump and Link
The encoding is via a special bit in the OPCODE_JUMP/BRANCH
instructions.

Called calljump and callbranch here, the assembler syntax is:

callbranch rN, label
calljump rN, rS

where rN will be loaded with the return address and rS is the
address of the subroutine.

This works by propogating the address of the instruction following
the current one down the pipeline, which is written into rN using
the normal methods

These are obviously all conditional, like other branch/jump
opcodes.