Added a tile-based VGA display controller
The first "applicaiton" of this softcore is intended to be a boulderdash
game. 20x15 tiles, which in turn are x2 16x16 16 colour tiles, borrowed
from the NES version of this game. Display is the usual 640x48 @ 60Hz.

Memories used are:

map_ram.v: 32x32 (so screen will scroll) x 8 bit map RAM, dual ported.
Because the iCE40 (the basic target) does not have proper DP RAM, the
array has to be doubled up. Read from port a for the display, and R/W
on port b for the processor.

tile_rom.v: 16x64 tile data. There is space for 64 different tiles.

palette_rom.v: 16x16 palette data. Format of output is 4 bits per RGB
in the form of xRGB.