Initial board glue for my iCE40UP Development Board
Brings in all the other modules, in the main softcore dir and under vga/
and sets up pinning for my dev board.

This needs to be factored into a further module layer, with the basic
top level linkage to external FPGA pins only in this package. That
can wait.

The build Makefile does not work quite right yet; the top level Makefile
needs to be run by hand first. But that'll come.