Animate the gems
This involved adding the "missing" tile positions so that the tile index is
now of the form:

0bZZAATTTT

TTTT is the tile type, ie. wall, gem, boulder. There are extra images
for tiles and boulders (and some others) which is where the AA bits come
in. ZZ is stored in the map RAM but not used to lookup the image. This
will probably be used for something later.

This means lots of masking before checking if a tile is a gem etc.

So far only gems are animated. The boulder animation looks silly
(boulders "jiggling") so I wont bother with those.

Lots of nops. :(