Boulderdash: Bats!
Simple algo: bits 4 and 5 hold the direction the bat is moving in, and
when it detects an obstacle it rotates instead of moving forward. Bit 7
is used as a toggle flag to determine if this bat has been moved in this
update, that way each bat is only moved once per update.

Bats can't kill the player yet, and are infact easily squashed just by
moving into them, just like any other tile which isn't a wall or boulder.