Add a program that writes a Boulderdash level to the I2C EEPROM
This involves a fair amount of twiddling as addresses have different
offsets on screen to in the EEPROM, and different tile sizes etc.

Intoduced two new ALU operations for shifting a register left and
right by a whole byte. This was useful in this code because of
managing the EEPROM address, which is spread over two bytes on the
wire and must be written through to the I2C controller as two
bytes.

Levels are 32x32 bytes (1KB) so the AT24C64 on my dev board can hold
8 levels.

TODO: on bytes which are not the last out in a transaction, there is
a one CPU clock wide glitch on the SDA line. This is harmless, but needs
removing.