Encapsulate all the VGA stuff in its own module
This contains: sync generator, map RAM, tile ROM and palette lookup.
It exposes the map RAM and the scrolling register to the processor.

This is hopefully a lot tidier and will make it easier to wrap the
modules in glue for a different FPGA board.

Also renamed the vga_clk external pin to vga_clock while I was here.