Started a status row for the display
This is a non scrolling row which will show the remaining lives and gems
that need to be collected before the exit will open. This is implemented
with a 32byte RAM. To save resources, it is not readable from the processor.

This involved a new tilemap, with numbers and tiles with a line at the
bottom which marks the border of the status line. Tiles have been
shuffled around, to keep the 0-9 tiles in a consequitive sequence.

After struggling with Gimp, I did the pixel painting on my Amiga using
Delux Paint. :)

Unfortunately without manually redoing it, the level I've created up
till now has been lost in the Tiled program I've been using since it
doesn't seem to cope well with changing tilesets. Nevermind. This will
probably not be the first level.

Next up will be making the status line show some data. This will involve
introducing a data-block in program memory which will hold things like the
player's starting position for that level, the number of gems that need
to be collected, and where the exit should open up.