Wrap the tone generator in an interface
This matches the I2C master and PS/2 controller. The other purpose was to
add a register for the tone generator that indicates if a tone is playing.
Processor code can poll on this bit. (One day we'll have interrupts)

Also a bunch of other small changes along the way, mostly to do with
"period" being a variable used by tests.vh. This got in the way of using
"period" for the tone 1/frequency.

Noticed that the PS/2 and I2C controllers were using the external clock and
not the CPU clock, which was naughty. This had the affect of slowing down
the I2C clock, so tweak some values (which should be params) to get it back
to ~100KHz. It would be nice to go faster, but the datasheet is fairly clear:
400KHz is only possible when the AT24C64 is run at 5V.

Tweaked the game starting so the user can't see the levels load, since at
100KHz they take about 0.1 seconds.