Added 2 more levels
Made the levels.txt be created by combing the other levels into one file.

The "parent" files are geneated with the Tiled program. I might check
those in at some point.

I was confused because all the levels looked the same after programing the
eeprom. It took a while until I noticed the eeporm writer was masking out the
level number when setting the level number, which was done it was reading from
the VGA map and I wanted it to write the same level again and again.

The level .txt files are symlinks for level 5 to 8, until I design real levels
for those slots.