Added support for the Terasic DE2-115
This is another "board" which, unforunately, relies on Quartus.
Specifiically this is tested on version 23.1, the last to support
the Cyclone IV, which is on this board.

This board is fantastic, with everything you'd ever want from an
FPGA board pretty much. It even has an I2C EEPROM, though it is only
4KByte in size. The only thing missing is a buzzer, so the boulderdash
game is rather silent. fMax is reported as about 33MHz, vs about 11MHz
on my iCE40UP5 board.

TODO: Address decoding needs to be moved out of the board dirs.

TODO: This board is actually called the DE2-115, not -110. Whoops!