ice40hxdevboard: add an initial board definition
Board isn't fully designed yet, but to prove the pinning is hopefully
correct I thought I'd do a port of this project to it.

The new board will have a ICE40HX4 in QFP144, SDRAM and a large
expansion slot for playing with other hardware, hopefully including
HDMI.

Current pinning constraints are pre any back-annotation to choose
the best connections, but results look good so far and it builds!