Finished an inital port for the ice40hx devboard, including HDMI
Renamed vga to video. Video timing is exactly the same as per VGA, so
the video module is the same on all boards. However the HX requires
a new DE (Data Enable) which is set when the beam is in the visible
portion of the screen. This is generated from the h_visible and
v_visible signals.

The biggest changes revolve are cutting the boulderdash tiles from
64 to 32. This is to make room for more code, which was needed
because of the new I2C write which is required by the HDMI controller,
a TFP410. This had the downside of needing to drop some tiles used
to animate the exit. But on the upside there are about 450 instruction
words available now to make the game more complete, if I wanted to do
that.