Added a UART and a test program
The UART is based on the one used in the MAXI030 FPGA, but obviously
rewritten in Verilog and tweaked quite a bit. Like most of the other
peripherals there is a _interface.v for it, which is responsible for
setting and clearing status bits and formatting up the status register
based on those bits.

The UART also has a trivial testbench, but it is currently only used
for generating a simple waveform, which exersises the UART after wiring
it for loopback, so you can see a byte get sent through the UART and
then received by it.

Added a basic test program to exersise the UART too; just a simple get
line then echo it back test. Soon it will be a full fledged machinecode
monitor!

Also plumbed the RGB LED back in. There is no led_interface as the
"peripheral" is completely trivial, but it's been tested with processor
code and works fine.

All this is on for the new fangled HX board. The other two boards will
probably notb uild at the moment.