Added a barrel shifter to the ALU
The logic/arith left/right shifts have been extended to be two
operand operations. The second operand is the shift count
(modulo 32). This replaces the byte shifting previously added
and allows shifts of any count. Surprisingly it does not seem
to impact the fMax a great deal. As the shifting amount comes
from the second operand, a register could be used to hold the
shift count if desired.

Coding is using Verilog bit shifting (<<, >> and >>> (arith))
operations. It would be interesting to see how this implemented
in the FPGA.

Tests have been extended very slightly, but more shifting tests
with different shifting amounts should be added.