Added initial go at a GitHub CI build
The build needs to include customasm, which is not packaged in any
debian (etc) repo, so it needs to be built. Otherwise this was
fairly easy to get going.

Added a bomb out to the maxicore32_tb run loop so it can't get stuck
in a loop, which would be sad for GitHub. Build and run log for "make
tests" is uploaded as an artifact.

Of course the tests are pretty lame and are mostly used for making
waveforms, which are usless here. Still, this is some kind of progress
to one day having a fully automated test suite.